Updated pause menu and keybinds
The pause menu now looks better and has a working button to go back to
the main menu.
The State's keybinds now use a struct with the keycode and a boolean to
keep track of the last pressed status, so the user can now press (or
release) a button and call an action once.