Added precompiled headers
The folder Source/Precompiled contains a header file with a batch of
headers, made of the most used ones.
The file headers.hpp.gch is the precompiled version of that batch and
can be used instead of the .hpp file to make the compilation faster.

Updated includes in project files to make use of this update.

Some other fixes.